Reorganize DataStore stuff into its own package
