+++ source/com/gmt2001/datastore/IniStore.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.gmt2001;
+package com.gmt2001.datastore;
 
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
